Fix a couple of missing includes for MSVC2005 and a C99 issue. Also, starting with 2.14.0, GLib won't provide a pipe(...
Original commit message from CVS:
Patch by: Ole André Vadla Ravnås  <ole.andre.ravnas@tandberg.com>
* gst/gsttrace.c:
* libs/gst/net/gstnetclientclock.c: (gst_net_client_clock_new):
* libs/gst/net/gstnettimepacket.c: (gst_net_time_packet_send):
* libs/gst/net/gstnettimeprovider.c: (gst_net_time_provider_new):
Fix a couple of missing includes for MSVC2005 and a C99 issue. Also,
starting with 2.14.0, GLib won't provide a pipe() macro any longer,
so use _pipe() directly (#492077).
* win32/common/dirent.c: (_treaddir):
Add a couple of casts to make it build without warnings with MSVC.
* win32/common/libgstreamer.def:
Add some more symbols that need to be exported.

+++ win32/common/dirent.c
@@ -285,7 +285,7 @@ _treaddir (_TDIR * dirp)
 
     /* Start the search */
 
-    dirp->dd_handle = _tfindfirst (dirp->dd_name, &(dirp->dd_dta));
+    dirp->dd_handle = (long) _tfindfirst (dirp->dd_name, &(dirp->dd_dta));
 
 
     if (dirp->dd_handle == -1) {
@@ -353,7 +353,7 @@ _treaddir (_TDIR * dirp)
 
      * file name. */
 
-    dirp->dd_dir.d_namlen = _tcslen (dirp->dd_dta.name);
+    dirp->dd_dir.d_namlen = (unsigned short) _tcslen (dirp->dd_dta.name);
 
     _tcscpy (dirp->dd_dir.d_name, dirp->dd_dta.name);
